Full rewrite to make compilation synchronous
The old code wasn't the easiest to work with due to a lot of async loops and potential race conditions. Since complation never occurs in a live running service there's no reason not to use synchronous (and therefore blocking) code.

Additionally there are places in other modules where we want to compile a directory of translations into a single object (see https://github.com/UKHomeOfficeForms/hof-template-partials/blob/master/translations/index.js for example). Since this uses the exact same approach as used here, it makes sense to make that a reusable module.